+++ android/src/main/java/io/sentry/react/RNSentryModule.java
@@ -269,6 +269,11 @@ public class RNSentryModule extends ReactContextBaseJavaModule {
                     }
                 }
 
+                if (totalFrames == 0 && slowFrames == 0 && frozenFrames == 0) {
+                    promise.resolve(null);
+                    return;
+                }
+
                 WritableMap map = Arguments.createMap();
                 map.putInt("totalFrames", totalFrames);
                 map.putInt("slowFrames", slowFrames);