+++ src/js/integrations/devicecontext.ts
@@ -1,5 +1,5 @@
 import { addGlobalEventProcessor, getCurrentHub } from "@sentry/core";
-import { Event, Integration } from "@sentry/types";
+import { Contexts, Event, Integration } from "@sentry/types";
 import { logger } from "@sentry/utils";
 
 import { NATIVE } from "../wrapper";
@@ -27,8 +27,16 @@ export class DeviceContext implements Integration {
       }
 
       try {
-        const deviceContexts = await NATIVE.fetchNativeDeviceContexts();
-        event.contexts = { ...deviceContexts, ...event.contexts };
+        const contexts = await NATIVE.fetchNativeDeviceContexts();
+
+        const context = contexts['context'] as Contexts ?? {};
+        const user = contexts['user'] ?? {};
+
+        event.contexts = { ...context, ...event.contexts };
+
+        if (!event.user) {
+          event.user = { ...user };
+        }
       } catch (e) {
         logger.log(`Failed to get device context from native: ${e}`);
       }