+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- fix(iOS): Bump Sentry Cocoa to 7.10.1 and report slow and frozen measurements (#)
+
 ## 3.3.1
 
 - feat: Support setting maxCacheItems #2102