+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- fix: Do not report empty measurements #1983
+
 ## 3.3.1
 
 - feat: Support setting maxCacheItems #2102