+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+
+- fix: Do not report empty measurements #1983
 - fix(iOS): Bump Sentry Cocoa to 7.10.1 and report slow and frozen measurements (#2132)
 
 ## 3.3.1